landmask tests: remove obsolete basemap category for tests
